# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AIS Catering",
        "contact": {},
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/caterings/{id}/dishes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Returns list of dishes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Category ID",
                        "name": "categoryID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of dishes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DishRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Add dish for certain category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "dish object",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/AddDishRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "dish object",
                        "schema": {
                            "$ref": "#/definitions/DishRequest"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/caterings/{id}/dishes/{dishId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Returns dishes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dish ID",
                        "name": "dishId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of dishes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DishRequest"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Returns 204 if success and 4xx error if failed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dish ID",
                        "name": "dishId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dish object",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/AddDishRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully updated"
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Soft delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dish ID",
                        "name": "dishId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/caterings/{id}/dishes/{dishId}/images": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Add image for certain dish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Dish ID",
                        "name": "dishId",
                        "in": "query"
                    },
                    {
                        "type": "file",
                        "description": "Image File",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "id of default image",
                        "name": "id",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "default image",
                        "schema": {
                            "$ref": "#/definitions/ImageResponse"
                        }
                    },
                    "201": {
                        "description": "Image model",
                        "schema": {
                            "$ref": "#/definitions/ImageResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/caterings/{id}/dishes/{dishId}/images/{imageId}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Updates image for certain dish",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "imageId",
                        "in": "path"
                    },
                    {
                        "type": "string",
                        "description": "Dish ID",
                        "name": "dishId",
                        "in": "path"
                    },
                    {
                        "type": "file",
                        "description": "Image File",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "default image",
                        "schema": {
                            "$ref": "#/definitions/ImageResponse"
                        }
                    },
                    "201": {
                        "description": "Image model",
                        "schema": {
                            "$ref": "#/definitions/ImageResponse"
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering dishes"
                ],
                "summary": "Soft delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Catering ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Image ID",
                        "name": "imageId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Dish ID",
                        "name": "dishId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/images": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catering images"
                ],
                "summary": "Returns list of images",
                "responses": {
                    "200": {
                        "description": "List of images",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/GetImagesResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AddDishRequest": {
            "type": "object",
            "required": [
                "categoryId",
                "name",
                "price",
                "weight"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "desc": {
                    "type": "string",
                    "example": "Очень вкусный"
                },
                "name": {
                    "type": "string",
                    "example": "грибной суп"
                },
                "price": {
                    "type": "number",
                    "example": 120
                },
                "weight": {
                    "type": "number",
                    "example": 250
                }
            }
        },
...
        "DishRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "weight"
            ],
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Image"
                    }
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "GetImagesResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "Image": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "ImageResponse": {
            "type": "object",
            "required": [
                "path"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
    }
}